#1557 making room for string -> enum mapping
