+++ processor/src/test/java/org/mapstruct/ap/test/value/enum2enum/OrderDto.java
@@ -3,8 +3,9 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.value;
+package org.mapstruct.ap.test.value.enum2enum;
 
+import org.mapstruct.ap.test.value.ExternalOrderType;
 
 /**
  * @author Gunnar Morling
