+++ processor/src/test/resources/fixtures/org/mapstruct/ap/test/value/enum2enum/OrderMapperImpl.java
@@ -3,16 +3,18 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.value;
+package org.mapstruct.ap.test.value.enum2enum;
 
 import javax.annotation.Generated;
+import org.mapstruct.ap.test.value.ExternalOrderType;
+import org.mapstruct.ap.test.value.OrderType;
 
 @Generated(
     value = "org.mapstruct.ap.MappingProcessor",
     date = "2017-02-20T21:25:45+0100",
     comments = "version: , compiler: javac, environment: Java 1.8.0_112 (Oracle Corporation)"
 )
-public class SpecialOrderMapperImpl implements SpecialOrderMapper {
+public class OrderMapperImpl implements OrderMapper {
 
     @Override
     public OrderDto orderEntityToDto(OrderEntity order) {
@@ -30,19 +32,23 @@ public class SpecialOrderMapperImpl implements SpecialOrderMapper {
     @Override
     public ExternalOrderType orderTypeToExternalOrderType(OrderType orderType) {
         if ( orderType == null ) {
-            return ExternalOrderType.DEFAULT;
+            return null;
         }
 
         ExternalOrderType externalOrderType;
 
         switch ( orderType ) {
-            case STANDARD: externalOrderType = null;
+            case EXTRA: externalOrderType = ExternalOrderType.SPECIAL;
+            break;
+            case STANDARD: externalOrderType = ExternalOrderType.DEFAULT;
+            break;
+            case NORMAL: externalOrderType = ExternalOrderType.DEFAULT;
             break;
             case RETAIL: externalOrderType = ExternalOrderType.RETAIL;
             break;
             case B2B: externalOrderType = ExternalOrderType.B2B;
             break;
-            default: externalOrderType = ExternalOrderType.SPECIAL;
+            default: throw new IllegalArgumentException( "Unexpected enum constant: " + orderType );
         }
 
         return externalOrderType;
@@ -51,7 +57,7 @@ public class SpecialOrderMapperImpl implements SpecialOrderMapper {
     @Override
     public OrderType externalOrderTypeToOrderType(ExternalOrderType orderType) {
         if ( orderType == null ) {
-            return OrderType.STANDARD;
+            return null;
         }
 
         OrderType orderType1;
@@ -59,7 +65,7 @@ public class SpecialOrderMapperImpl implements SpecialOrderMapper {
         switch ( orderType ) {
             case SPECIAL: orderType1 = OrderType.EXTRA;
             break;
-            case DEFAULT: orderType1 = null;
+            case DEFAULT: orderType1 = OrderType.STANDARD;
             break;
             case RETAIL: orderType1 = OrderType.RETAIL;
             break;
@@ -69,26 +75,5 @@ public class SpecialOrderMapperImpl implements SpecialOrderMapper {
         }
 
         return orderType1;
-    }
-
-    @Override
-    public ExternalOrderType anyRemainingToNull(OrderType orderType) {
-        if ( orderType == null ) {
-            return ExternalOrderType.DEFAULT;
-        }
-
-        ExternalOrderType externalOrderType;
-
-        switch ( orderType ) {
-            case STANDARD: externalOrderType = null;
-            break;
-            case RETAIL: externalOrderType = ExternalOrderType.RETAIL;
-            break;
-            case B2B: externalOrderType = ExternalOrderType.B2B;
-            break;
-            default: externalOrderType = null;
-        }
-
-        return externalOrderType;
     }
 }
