+++ processor/src/test/resources/fixtures/org/mapstruct/ap/test/value/enum2enum/SpecialOrderMapperImpl.java
@@ -3,16 +3,18 @@
  *
  * Licensed under the Apache License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package org.mapstruct.ap.test.value;
+package org.mapstruct.ap.test.value.enum2enum;
 
 import javax.annotation.Generated;
+import org.mapstruct.ap.test.value.ExternalOrderType;
+import org.mapstruct.ap.test.value.OrderType;
 
 @Generated(
     value = "org.mapstruct.ap.MappingProcessor",
     date = "2017-02-20T21:25:45+0100",
     comments = "version: , compiler: javac, environment: Java 1.8.0_112 (Oracle Corporation)"
 )
-public class OrderMapperImpl implements OrderMapper {
+public class SpecialOrderMapperImpl implements SpecialOrderMapper {
 
     @Override
     public OrderDto orderEntityToDto(OrderEntity order) {
@@ -30,23 +32,19 @@ public class OrderMapperImpl implements OrderMapper {
     @Override
     public ExternalOrderType orderTypeToExternalOrderType(OrderType orderType) {
         if ( orderType == null ) {
-            return null;
+            return ExternalOrderType.DEFAULT;
         }
 
         ExternalOrderType externalOrderType;
 
         switch ( orderType ) {
-            case EXTRA: externalOrderType = ExternalOrderType.SPECIAL;
-            break;
-            case STANDARD: externalOrderType = ExternalOrderType.DEFAULT;
-            break;
-            case NORMAL: externalOrderType = ExternalOrderType.DEFAULT;
+            case STANDARD: externalOrderType = null;
             break;
             case RETAIL: externalOrderType = ExternalOrderType.RETAIL;
             break;
             case B2B: externalOrderType = ExternalOrderType.B2B;
             break;
-            default: throw new IllegalArgumentException( "Unexpected enum constant: " + orderType );
+            default: externalOrderType = ExternalOrderType.SPECIAL;
         }
 
         return externalOrderType;
@@ -55,7 +53,7 @@ public class OrderMapperImpl implements OrderMapper {
     @Override
     public OrderType externalOrderTypeToOrderType(ExternalOrderType orderType) {
         if ( orderType == null ) {
-            return null;
+            return OrderType.STANDARD;
         }
 
         OrderType orderType1;
@@ -63,7 +61,7 @@ public class OrderMapperImpl implements OrderMapper {
         switch ( orderType ) {
             case SPECIAL: orderType1 = OrderType.EXTRA;
             break;
-            case DEFAULT: orderType1 = OrderType.STANDARD;
+            case DEFAULT: orderType1 = null;
             break;
             case RETAIL: orderType1 = OrderType.RETAIL;
             break;
@@ -73,5 +71,26 @@ public class OrderMapperImpl implements OrderMapper {
         }
 
         return orderType1;
+    }
+
+    @Override
+    public ExternalOrderType anyRemainingToNull(OrderType orderType) {
+        if ( orderType == null ) {
+            return ExternalOrderType.DEFAULT;
+        }
+
+        ExternalOrderType externalOrderType;
+
+        switch ( orderType ) {
+            case STANDARD: externalOrderType = null;
+            break;
+            case RETAIL: externalOrderType = ExternalOrderType.RETAIL;
+            break;
+            case B2B: externalOrderType = ExternalOrderType.B2B;
+            break;
+            default: externalOrderType = null;
+        }
+
+        return externalOrderType;
     }
 }
